combine redux tests
